Bug fixing

+++ src/Scripting/Events/SceneColor.bpi.cpp
@@ -5,7 +5,7 @@
 WRAP(EventSceneColor)
 {
 	class_<Event::SceneColor>("EventSceneColor", no_init)
-		.def(init<uint, uint, const Vec3&>())
+		.def(init<float, float, const Vec3&>())
 		.def(init<const Event::SceneColor&>())
 	;
 }
